requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
asyncpg
python-jose
httpx
google-generativeai
google